refactor(verifier-v2): move length proofs to kzg/committer
length proof don't require any of the SRS or kzgConfig stuff in the verifier. They are simple functions that can just remain that. No need to be methods on any struct.

Also felt it was more appropriate for them to be next to where blob length commitments/proof are generated, since their job is to verify those. So moved them to kzg/committer along with their tests. This cleans the Verifier struct to have a single purpose: verify frames.
