+++ encoding/kzg/verifier/v2/verifier.go
@@ -4,13 +4,9 @@ import (
 	"errors"
 	"fmt"
 	gomath "math"
-	"math/bits"
 	"sync"
 
-	"github.com/Layr-Labs/eigenda/common/math"
 	"github.com/Layr-Labs/eigenda/encoding"
-	"github.com/Layr-Labs/eigenda/resources/srs"
-
 	"github.com/Layr-Labs/eigenda/encoding/fft"
 	"github.com/Layr-Labs/eigenda/encoding/kzg"
 	"github.com/Layr-Labs/eigenda/encoding/rs"
@@ -89,57 +85,9 @@ func (v *Verifier) newKzgVerifier(params encoding.EncodingParams) (*Parametrized
 	fs := fft.NewFFTSettings(n)
 
 	return &ParametrizedVerifier{
-		KzgConfig: v.kzgConfig,
-		g1SRS:     v.G1SRS,
-		Fs:        fs,
+		g1SRS: v.G1SRS,
+		Fs:    fs,
 	}, nil
-}
-
-func (v *Verifier) VerifyBlobLength(commitments encoding.BlobCommitments) error {
-	return v.VerifyLengthProof(
-		(*bn254.G2Affine)(commitments.LengthCommitment),
-		(*bn254.G2Affine)(commitments.LengthProof),
-		uint64(commitments.Length),
-	)
-}
-
-// VerifyLengthProof verifies the length proof (low degree proof).
-// See https://layr-labs.github.io/eigenda/protocol/architecture/encoding.html#validation-via-kzg
-// Since it doesn't depend on the encoding parameters, we leave it as a method of Verifier, not ParametrizedVerifier.
-func (v *Verifier) VerifyLengthProof(
-	lengthCommit *bn254.G2Affine, lengthProof *bn254.G2Affine, commitmentLength uint64,
-) error {
-	// This also prevents commitmentLength=0.
-	if !math.IsPowerOfTwo(commitmentLength) {
-		return fmt.Errorf("commitment length %d is not a power of 2", commitmentLength)
-	}
-	// Because commitmentLength is power of 2, we know its represented as 100..0 in binary,
-	// so counting the number of trailing zeros gives us log2(commitmentLength).
-	// We need commitmentLengthLog <= 27 because we have hardcoded SRS points only for that range.
-	commitmentLengthLog := bits.TrailingZeros64(commitmentLength)
-	if commitmentLengthLog > 27 {
-		return fmt.Errorf("commitment length %d is > max possible 2^28", commitmentLength)
-	}
-	// g1Challenge = [tau^(2^28 - commitmentLength)]_1
-	// G1ReversePowerOf2SRS contains the 28 hardcoded points that we need.
-	g1Challenge := srs.G1ReversePowerOf2SRS[commitmentLengthLog]
-
-	err := verifyLengthProof(lengthCommit, lengthProof, &g1Challenge)
-	if err != nil {
-		return fmt.Errorf("low degree proof: %w", err)
-	}
-	return nil
-}
-
-// This function verifies a low degree proof against a poly commitment.
-// We wish to show x^shift poly = shiftedPoly, with shift = 2^28 - blob_length.
-// We verify this by checking the pairing equation:
-// e( s^shift G1, p(s)G2 ) = e( G1, p(s^shift)G2 )
-// Note that we also need to verify that the blob_commitment and length_commitment are equivalent,
-// by verifying the other pairing equation: e(blob_commitment,G2) = e(length_commitment,C2)
-// TODO(samlaf): can we move that other pairing check in here?
-func verifyLengthProof(lengthCommit *bn254.G2Affine, proof *bn254.G2Affine, g1Challenge *bn254.G1Affine) error {
-	return pairingsVerify(g1Challenge, lengthCommit, &kzg.GenG1, proof)
 }
 
 // VerifyFrame verifies a single frame against a commitment.
@@ -197,22 +145,4 @@ func toUint64Array(chunkIndices []encoding.ChunkNumber) []uint64 {
 		res[i] = uint64(d)
 	}
 	return res
-}
-
-func pairingsVerify(a1 *bn254.G1Affine, a2 *bn254.G2Affine, b1 *bn254.G1Affine, b2 *bn254.G2Affine) error {
-	var negB1 bn254.G1Affine
-	negB1.Neg(b1)
-
-	P := [2]bn254.G1Affine{*a1, negB1}
-	Q := [2]bn254.G2Affine{*a2, *b2}
-
-	ok, err := bn254.PairingCheck(P[:], Q[:])
-	if err != nil {
-		return fmt.Errorf("PairingCheck: %w", err)
-	}
-	if !ok {
-		return errors.New("PairingCheck pairing not ok. SRS is invalid")
-	}
-
-	return nil
 }
